fix: removde deprecated controller methods
