+++ python/src/controller.py
@@ -1,15 +1,11 @@
 import time
-import mqttbroker as mqtt
 import databasemanager as dbm
-import threading
-          
+
 class Controller:
     def save(data):
         if data == None:
             return False
-        
-        #Todo: deserialization of sensor data
-        
+        #Todo: adjust model
         self.db_manager.persist(Temperature(temperature=data, sensor_id=data))
     
     def load(queryset=dict):
@@ -18,42 +14,10 @@ class Controller:
         self.db_manager.load(Temperature, queryset)
         return True
     
-    def update(primary_key, change_key, change_value):
+    def update(self, primary_key, change_key, change_value):
+        #ToDo: Remove timestamp generation of deprecated
         temperature_object = self.load({"timestamp": primary_key})
         self.db_manager.update(temperature_object, change_key, change_value)
     
-    def stop_mqtt():
-        self.mqtt_running = False
-        
-    
-    def start_mqtt(self):    
-        connected = False
-        try:
-            client = self.mqtt_broker.connect_mqtt()
-
-            topic = "topic_name or something"
-            self.mqtt_broker.subscribe(topic, client)
-            connected = True
-        except Exception as e:
-            connected = False
-            print(e)
-
-        if connected:
-            broker_thread = threading.Thread(target=client.loop_forever)
-
-            broker_thread.start()
-
-            self.mqtt_running = True
-            print("I work")
-            while self.mqtt_running:
-                if self.mqtt_broker.queue:
-                    payload = self.mqtt_broker.queue.pop(0)
-                    print(payload)
-            else:
-                broker_thread._stop()
-    
     def __init__(self):
-        print("I've been initialized")
         self.db_manager = dbm.DatabaseManager()
-        self.mqtt_broker = mqtt.Broker()
-        self.mqtt_running = False
