+++ python/src/mqttbroker.py
@@ -25,6 +25,36 @@ class Broker:
         client.on_connect = on_connect
         client.connect(broker, port)
         return client
+    
+    def stop_mqtt(self):
+        self.mqtt_running = False
+        
+    # ToDo: fix this method
+    def start_mqtt(self):    
+        connected = False
+        try:
+            client = self.mqtt_broker.connect_mqtt()
+
+            topic = "topic_name or something"
+            self.mqtt_broker.subscribe(topic, client)
+            connected = True
+        except Exception as e:
+            connected = False
+            print(e)
+
+        if connected:
+            broker_thread = threading.Thread(target=client.loop_forever)
+
+            broker_thread.start()
+
+            self.mqtt_running = True
+            print("I work")
+            while self.mqtt_running:
+                if self.mqtt_broker.queue:
+                    payload = self.mqtt_broker.queue.pop(0)
+                    print(payload)
+            else:
+                broker_thread._stop()
 
     def on_disconnect(self, client, userdata, rc):
         logging.info("Disconnected with result code: %s", rc)
